Remove a race from the s3c2410 touch driver
There seems a race in the driver when it uses the bit 15 from the dat0 and dat1
register: These bits are only valid when the pen interrupt feature is enabled.
This is not the case when a regular touchscreen X/Y conversion is running. It
only works due to a small race between s3c24xx_ts_select(), stylus_irq() and
touch_timer_fire(). It is broken immediately when the debug output of the
touchscreen driver will be enabled or the debug output of the ADC driver. In
this case the conversion never stops, even there is no pressure on the touch
anymore.

This patch simplifies the driver and stops any further conversion if the pen up
interrupt is received. Pen up and down detection is now done only in the
pen interrupt routine. This also prevents the driver forwarding garbage data
to userland, because when the pen is up, the X value is always 0.

Note: This patch makes the driver stop working on a S3C2410 CPU.

Singed-off-by: Juergen Beisert <jbe@pengutronix.de>