ARC: [haps] Add Virtio support
For now we add 5 Virtio slots in .dts and enable block and
network devices via Virtio-MMIO.

Note we also enable DEVTMPFS automount for more convenient use
of external root file-stystem. Before that we used to use built-in
Initramfs which didn't automount DEVTMPFS anyways so we didn't need
that option, while now it starts making sense.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>